Migration from fanstatic to webassets + minor six changes.
