+++ ckanext/dhis2harvester/assets/pivot_tables.js
