+++ ckanext/dhis2harvester/assets/tabs.js
