+++ ckanext/dhis2harvester/assets/toggle_box.js
