+++ ckanext/dhis2harvester/templates/source/dhis2_credentials.html
@@ -1,5 +1,4 @@
 {% import 'macros/form.html' as form %}
-{% resource 'dhis2harvester/toggle_box.js' %}
 {% block visibility %}
 <div class="pivot-table-new-1 {% if not data.action or data.action != 'pivot_table_new_1' %}hidden{% endif %}">
 {% endblock %}
