+++ ckanext/dhis2harvester/templates/source/pivot_table_base.html
@@ -1,8 +1,13 @@
 {% extends "source/admin_base.html" %}
 
 {% block styles %}
-{{super()}}
-<link rel="stylesheet" type="text/css" href="{{ h.url_for_static('/dhis2_harvester_source_form.css')}}">
+  {{ super() }}
+  {% asset 'ckanext-dhis2harvester/main-css' %}
+{% endblock %}
+
+{% block scripts %}
+  {{ super() }}
+  {% asset 'ckanext-dhis2harvester/main' %}
 {% endblock %}
 
 {% block breadcrumb_content %}
