+++ ckanext/dhis2harvester/templates/source/pivot_table_new.html
@@ -6,11 +6,9 @@
 {% else %}
   {% extends "source/pivot_table_base.html" %}
   {% import 'macros/form.html' as form %}
-  {% resource 'harvest-extra-field/main' %}
 
 
   {% block form_content %}
-  {% resource 'dhis2harvester/throbber.js' %}
 
   <form id="source-new" class="dhis2-source-new form-horizontal dataset-form {{ h.bootstrap_version() }}" method="post">
