+++ ckanext/dhis2harvester/templates/source/pivot_table_new_2.html
@@ -1,5 +1,4 @@
 {% import 'macros/form.html' as form %}
-{% resource 'dhis2harvester/pivot_tables.js' %}
 
 
 <div class="pivot-table-new-2 {% if data.action != 'pivot_table_new_2' %}hidden{% endif %}">
