+++ ckanext/dhis2harvester/templates/source/pivot_table_refresh.html
@@ -1,6 +1,5 @@
 {% extends "source/pivot_table_base.html" %}
 {% import 'macros/form.html' as form %}
-{% resource 'harvest-extra-field/main' %}
 
 {% block subtitle %}{{ _('Fetch DHIS2 Pivot Table') }}{% endblock %}
